[#243] FEAT: AddPhotoScrollView생성 및 작성
- 스택뷰 생성
- 스택뷰와 스크롤뷰의 크기 맞춤
- 좌우 스크롤바 가리기
